% Generated by roxygen2 (4.0.0): do not edit by hand
\name{turtle_do}
\alias{turtle_do}
\title{Evaluate a Turtle expresion}
\usage{
turtle_do(expr)
}
\arguments{
\item{expr}{expresion to evaluate}
}
\description{
\code{turtle_do} evaluate a given expression
}
\details{
The terrarium must be initialized prior to using
these functions, see \code{\link{turtle_init}}.

\code{turtle_do} evaluate a given expresion with hidden Turtle, which
increase the rate in case of complicated ...
}
\examples{
turtle_init()
turtle_do(expr = {
   turtle_move(100)
   turtle_turn(45)
   turtle_move(100)
})
}
\seealso{
Other TurtleGraphics: \code{\link{TurtleGraphics-package}};
  \code{\link{turtle_backward}},
  \code{\link{turtle_forward}}, \code{\link{turtle_move}};
  \code{\link{turtle_col}}, \code{\link{turtle_lty}},
  \code{\link{turtle_lwd}}, \code{\link{turtle_param}};
  \code{\link{turtle_getangle}},
  \code{\link{turtle_getpos}}; \code{\link{turtle_goto}},
  \code{\link{turtle_setangle}},
  \code{\link{turtle_setpos}}; \code{\link{turtle_hide}},
  \code{\link{turtle_show}}; \code{\link{turtle_init}};
  \code{\link{turtle_left}}, \code{\link{turtle_right}},
  \code{\link{turtle_turn}}; \code{\link{turtle_reset}};
  \code{\link{turtle_status}}
}

