\name{turtle_getpos}
\alias{turtle_getangle}
\alias{turtle_getpos}
\title{t.b.d.}
\usage{
turtle_getpos()

turtle_getangle()
}
\description{
t.b.d.
}
\details{
t.b.d.
}
\seealso{
Other TurtleGraphics: \code{\link{TurtleGraphics-package}};
\code{\link{turtle_backward}},
\code{\link{turtle_forward}}, \code{\link{turtle_move}};
\code{\link{turtle_col}}, \code{\link{turtle_lty}},
\code{\link{turtle_lwd}}, \code{\link{turtle_param}};
\code{\link{turtle_goto}}, \code{\link{turtle_setangle}},
\code{\link{turtle_setpos}}; \code{\link{turtle_hide}},
\code{\link{turtle_show}}; \code{\link{turtle_init}};
\code{\link{turtle_left}}, \code{\link{turtle_right}},
\code{\link{turtle_turn}}; \code{\link{turtle_reset}};
\code{\link{turtle_status}}
}

