% Generated by roxygen2 (4.0.0): do not edit by hand
\name{turtle_goto}
\alias{turtle_goto}
\alias{turtle_setangle}
\alias{turtle_setpos}
\title{Set a Turtle position and direction}
\usage{
turtle_goto(x, y)

turtle_setpos(x, y)

turtle_setangle(angle)
}
\arguments{
\item{x,y}{coordinates specifying new Turtle location}

\item{angle}{rotation angle in degrees}
}
\description{
Functions \code{turtle_goto} and \code{turtle_setpos} moves a Turtle to a
given location without changing his direction, while \code{turtle_setangle}
function rotate a Turtle with a given angle.
}
\details{
The turtle must be initialized prior to using
\code{turtle_goto}, \code{turtle_setpos} and \code{turtle_setangle}
, see \code{\link{turtle_init}}. Moreover, when given location
(x, y) lies outside the terrarium, the bahavior of these functions depends
on choosen parameter \code{mode}.

The \code{turtle_goto} may draw or not the path between current Turtle
position and a new location which depends on a user settings,
see \code{\link{turtle_up}}, \code{\link{turtle_down}}. In a case of
\code{turtle_setpos}, however, the path drawing is always unable.
}
\seealso{
Other TurtleGraphics: \code{\link{TurtleGraphics-package}};
  \code{\link{turtle_backward}},
  \code{\link{turtle_forward}}, \code{\link{turtle_move}};
  \code{\link{turtle_col}}, \code{\link{turtle_lty}},
  \code{\link{turtle_lwd}}, \code{\link{turtle_param}};
  \code{\link{turtle_getangle}},
  \code{\link{turtle_getpos}}; \code{\link{turtle_hide}},
  \code{\link{turtle_show}}; \code{\link{turtle_init}};
  \code{\link{turtle_left}}, \code{\link{turtle_right}},
  \code{\link{turtle_turn}}; \code{\link{turtle_reset}};
  \code{\link{turtle_status}}
}

