\name{turtle_init}
\alias{turtle_init}
\title{Initialization of the Turtle Program}
\usage{
turtle_init(size = 1000)
}
\arguments{
  \item{size}{A non-negative scalar denoting a number of
  turtle moves to be drawn}
}
\description{
\code{turtle_init()} create an empty page with the Turtle
image displayed in the middle.
}

