\name{turtle_status}
\alias{turtle_status}
\title{Turtle Position Status}
\usage{
turtle_status()
}
\description{
\code{turtle_status()} gives information about current
Turtle position and values of graphical parameters.
}
\details{
\code{turtle_status()} returns two element list describing
graphical parameters valus and Turtle current position.
}
\seealso{
Other TurtleGraphics: \code{\link{TurtleGraphics-package}};
\code{\link{turtle_backward}},
\code{\link{turtle_backward}},
\code{\link{turtle_forward}}, \code{\link{turtle_forward}},
\code{\link{turtle_move}}, \code{\link{turtle_move}};
\code{\link{turtle_hide}}, \code{\link{turtle_show}};
\code{\link{turtle_left}}, \code{\link{turtle_left}},
\code{\link{turtle_right}}, \code{\link{turtle_right}},
\code{\link{turtle_turn}}, \code{\link{turtle_turn}};
\code{\link{turtle_reset}}
}

