% Generated by roxygen2 (4.0.0): do not edit by hand
\name{turtle_up}
\alias{turtle_down}
\alias{turtle_up}
\title{Turn on or off Turtle Trace Drawing}
\usage{
turtle_up()

turtle_down()
}
\description{
When the Turtle moves, it may or may not
leave a visible trace. These functions
control such a behavior.
}
\details{
The Turtle must be initialized prior to using
this function, see \code{\link{turtle_init}}.
}
\seealso{
Other TurtleGraphics: \code{\link{TurtleGraphics-package}};
  \code{\link{turtle_backward}},
  \code{\link{turtle_forward}}, \code{\link{turtle_move}};
  \code{\link{turtle_col}}, \code{\link{turtle_lty}},
  \code{\link{turtle_lwd}}, \code{\link{turtle_param}};
  \code{\link{turtle_do}}; \code{\link{turtle_getangle}},
  \code{\link{turtle_getpos}}; \code{\link{turtle_goto}},
  \code{\link{turtle_setangle}},
  \code{\link{turtle_setpos}}; \code{\link{turtle_hide}},
  \code{\link{turtle_show}}; \code{\link{turtle_init}};
  \code{\link{turtle_left}}, \code{\link{turtle_right}},
  \code{\link{turtle_turn}}; \code{\link{turtle_reset}};
  \code{\link{turtle_status}}
}

