\docType{package}
\name{TurtleGraphics-package}
\alias{TurtleGraphics-package}
\title{Turtle Graphics in R}
\description{
TO BE DONE
}
\details{
TO BE DONE
}
\author{
Anna Cena \email{cena@rexamine.com} [aut]\cr Marek
Gagolewski \email{gagolews@rexamine.com} [aut]\cr Marcin
Kosinski \email{m.p.kosinski@gmail.com} [aut]\cr Barbara
Zogala-Siudem \email{zogala@rexamine.com} [aut, cre]
}

