\name{hide}
\alias{hide}
\alias{hide_turtle}
\title{Hide Turtle Image}
\usage{
hide()

hide_turtle()
}
\description{
\code{hide()} function makes Turtle image invisible while
the path is drawn.
}
\details{
To use turn, the tutrle must be initiated, see
\code{\link{turtle_init}}. While Turtle has been hiden, the
\code{show()} function should be used to display the Turtle
image.
}
\seealso{
\code{\link{show}}
}

