% Generated by roxygen2 (4.0.0): do not edit by hand
\name{turtle_getpos}
\alias{turtle_getangle}
\alias{turtle_getpos}
\title{Get a Turtle position and direction}
\usage{
turtle_getpos()

turtle_getangle()
}
\description{
\code{turtle_getpos} returns a Turtle current position

\code{turtle_getangle} returns a Turtle current direction
}
\details{
The turtle must be initialized prior to using
\code{turtle_getpos} and \code{turtle_getangle},
see \code{\link{turtle_init}}.

The \code{turtle_getpos} returns a vector of length two which
specifying the \code{x} and \code{y} coordinates.
The \code{turtle_getangle} returns the the \code{angle} in which Turtle is rotated.
}
\seealso{
Other TurtleGraphics: \code{\link{TurtleGraphics-package}};
  \code{\link{turtle_backward}},
  \code{\link{turtle_forward}}, \code{\link{turtle_move}};
  \code{\link{turtle_col}}, \code{\link{turtle_lty}},
  \code{\link{turtle_lwd}}, \code{\link{turtle_param}};
  \code{\link{turtle_goto}}, \code{\link{turtle_setangle}},
  \code{\link{turtle_setpos}}; \code{\link{turtle_hide}},
  \code{\link{turtle_show}}; \code{\link{turtle_init}};
  \code{\link{turtle_left}}, \code{\link{turtle_right}},
  \code{\link{turtle_turn}}; \code{\link{turtle_reset}};
  \code{\link{turtle_status}}
}

