% Generated by roxygen2 (4.0.2): do not edit by hand
\name{turtle_getpos}
\alias{turtle_getangle}
\alias{turtle_getpos}
\title{Get the Turtle's Current Position and Direction}
\usage{
turtle_getpos()

turtle_getangle()
}
\value{
Both functions return a (named) numeric vector.
\code{turtle_getpos} returns a vector of length two which
specifies the \code{x} and \code{y} coordinates.
The \code{turtle_getangle} returns the angle.
}
\description{
\code{turtle_getpos} returns the Turtle's current position on the plane.

\code{turtle_getangle} returns the Turtle's current direction,
in degrees. An angle of 0 represents a north-facing Turtle.
}
\details{
The terrarium must be initialized prior to using these functions,
see \code{\link{turtle_init}}.
}
\examples{
turtle_init()
turtle_getpos()["x"] # x coordinate
turtle_getpos()["y"] # y coordinate
}
\seealso{
Other TurtleGraphics: \code{\link{TurtleGraphics-package}};
  \code{\link{turtle_backward}},
  \code{\link{turtle_forward}}, \code{\link{turtle_move}};
  \code{\link{turtle_col}}, \code{\link{turtle_lty}},
  \code{\link{turtle_lwd}}, \code{\link{turtle_param}};
  \code{\link{turtle_down}}, \code{\link{turtle_up}};
  \code{\link{turtle_do}}; \code{\link{turtle_goto}},
  \code{\link{turtle_setangle}},
  \code{\link{turtle_setpos}}; \code{\link{turtle_hide}},
  \code{\link{turtle_show}}; \code{\link{turtle_init}};
  \code{\link{turtle_left}}, \code{\link{turtle_right}},
  \code{\link{turtle_turn}}; \code{\link{turtle_reset}};
  \code{\link{turtle_status}}
}

