% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/goto.R
\name{turtle_goto}
\alias{turtle_goto}
\alias{turtle_setpos}
\alias{turtle_setangle}
\title{Set the Turtle's Position and Direction}
\usage{
turtle_goto(x, y)

turtle_setpos(x, y)

turtle_setangle(angle)
}
\arguments{
\item{x, y}{numeric; coordinates specifying new Turtle's location.}

\item{angle}{numeric; rotation angle in degrees.}
}
\description{
\code{turtle_goto} and \code{turtle_setpos} move the Turtle to a
given location without changing its direction.

\code{turtle_setangle}
rotates the Turtle to a given (absolute) angle,
where 0 denotes a north-facing Turtle.
}
\details{
The terrarium must be initialized prior to using
these functions, see \code{\link{turtle_init}}.

If the given location (x, y) lies outside the terrarium,
the behavior of these functions depends
on the \code{mode} argument in \code{\link{turtle_init}}.

\code{turtle_goto} may draw the path between the current Turtle's
position and the new location. Its behavior depends on the current
plot settings,
see \code{\link{turtle_up}}, \code{\link{turtle_down}}. In case of
\code{turtle_setpos}, however, the path drawing is always disabled.
}
\seealso{
Other TurtleGraphics: \code{\link{TurtleGraphics-package}},
  \code{\link{turtle_do}}, \code{\link{turtle_getpos}},
  \code{\link{turtle_init}}, \code{\link{turtle_move}},
  \code{\link{turtle_param}}, \code{\link{turtle_reset}},
  \code{\link{turtle_show}}, \code{\link{turtle_status}},
  \code{\link{turtle_turn}}, \code{\link{turtle_up}}
}
