\name{turtle_init}
\alias{turtle_init}
\title{Set Up a New, Shiny Terrarium}
\usage{
turtle_init()
}
\description{
This function creates a new empty plot with the turtle
centered on the board and facing to the north.
}
\details{
After the \code{turtle_init()} function was called you can
e.g. move the turtle with the \code{\link{turtle_forward}}
function, turn its direction with
\code{\link{turtle_right}} or set display parameters of the
turtle trace, see \code{\link{turtle_param}}.
}
\seealso{
Other TurtleGraphics: \code{\link{TurtleGraphics-package}};
\code{\link{turtle_backward}},
\code{\link{turtle_forward}}, \code{\link{turtle_move}};
\code{\link{turtle_col}}, \code{\link{turtle_down}},
\code{\link{turtle_draw}}, \code{\link{turtle_lty}},
\code{\link{turtle_lwd}}, \code{\link{turtle_param}},
\code{\link{turtle_up}}; \code{\link{turtle_hide}},
\code{\link{turtle_show}}; \code{\link{turtle_left}},
\code{\link{turtle_right}}, \code{\link{turtle_turn}};
\code{\link{turtle_reset}}; \code{\link{turtle_status}}
}

