\name{turtle_init}
\alias{turtle_init}
\title{Starting the Turtle Program}
\usage{
turtle_init()
}
\description{
\code{turtle_init()} function create an empty page with the
turtle image displayed in the center.
}
\details{
After \code{turtle_init()} was called you can move the
turtle with \code{\link{move_forward}} or
\code{\link{move_bakcward}} functions, turn its direction
with \code{\link{turn}} or set graphical parameters of the
turtle trace, see \code{\link{set_param}}.
}

