% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reset.R
\name{turtle_reset}
\alias{turtle_reset}
\title{Reset the Turtle's Position and Direction}
\usage{
turtle_reset()
}
\description{
This function resets the Turtle's position, direction,
and graphical options.
}
\details{
The Turtle must be initialized prior to using
this function, see \code{\link{turtle_init}}.

After a call to this function, the Turtle will be placed
in the terrarium's center and it will be directed
to the north.

The drawing remains unchanged.
}
\examples{
turtle_init()
turtle_forward(4)
turtle_param(col="red", lty=2, lwd=3)
turtle_reset()
turtle_left(45)
turtle_forward(3)

}
\seealso{
Other TurtleGraphics: \code{\link{TurtleGraphics-package}},
  \code{\link{turtle_do}}, \code{\link{turtle_getpos}},
  \code{\link{turtle_goto}}, \code{\link{turtle_init}},
  \code{\link{turtle_move}}, \code{\link{turtle_param}},
  \code{\link{turtle_show}}, \code{\link{turtle_status}},
  \code{\link{turtle_turn}}, \code{\link{turtle_up}}
}
