% Generated by roxygen2 (4.0.2): do not edit by hand
\name{turtle_reset}
\alias{turtle_reset}
\title{Reset the Turtle's Position and Direction}
\usage{
turtle_reset()
}
\description{
This function resets the Turtle's position, direction,
and graphical options.
}
\details{
The Turtle must be initialized prior to using
this function, see \code{\link{turtle_init}}.

After a call to this function, the Turtle will be placed
in the terrarium's center and it will be directed
to the north.

The drawing remains unchanged.
}
\examples{
turtle_init()
turtle_forward(4)
turtle_param(col="red", lty=2, lwd=3)
turtle_reset()
turtle_left(45)
turtle_forward(3)
}
\seealso{
Other TurtleGraphics: \code{\link{TurtleGraphics-package}};
  \code{\link{turtle_backward}},
  \code{\link{turtle_forward}}, \code{\link{turtle_move}};
  \code{\link{turtle_col}}, \code{\link{turtle_lty}},
  \code{\link{turtle_lwd}}, \code{\link{turtle_param}};
  \code{\link{turtle_down}}, \code{\link{turtle_up}};
  \code{\link{turtle_do}}; \code{\link{turtle_getangle}},
  \code{\link{turtle_getpos}}; \code{\link{turtle_goto}},
  \code{\link{turtle_setangle}},
  \code{\link{turtle_setpos}}; \code{\link{turtle_hide}},
  \code{\link{turtle_show}}; \code{\link{turtle_init}};
  \code{\link{turtle_left}}, \code{\link{turtle_right}},
  \code{\link{turtle_turn}}; \code{\link{turtle_status}}
}

