% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show.R
\name{turtle_show}
\alias{turtle_show}
\alias{turtle_hide}
\title{Show or Hide the Turtle}
\usage{
turtle_show()

turtle_hide()
}
\description{
These functions enable or disable displaying the Turtle's
image on the screen.
}
\details{
The Turtle must be initialized prior to using
this function, see \code{\link{turtle_init}}.

It is recommended to hide the Turtle when
performing multiple Turtle moves, for efficiency reasons,
see also \code{\link{turtle_do}}.
}
\examples{
turtle_init()
turtle_forward(4)
turtle_hide()
turtle_left(30)
turtle_forward(3)

}
\seealso{
Other TurtleGraphics: \code{\link{TurtleGraphics-package}},
  \code{\link{turtle_do}}, \code{\link{turtle_getpos}},
  \code{\link{turtle_goto}}, \code{\link{turtle_init}},
  \code{\link{turtle_move}}, \code{\link{turtle_param}},
  \code{\link{turtle_reset}}, \code{\link{turtle_status}},
  \code{\link{turtle_turn}}, \code{\link{turtle_up}}
}
