% Generated by roxygen2 (4.0.1): do not edit by hand
\name{turtle_show}
\alias{turtle_hide}
\alias{turtle_show}
\title{Show or Hide the Turtle}
\usage{
turtle_show()

turtle_hide()
}
\description{
These functions enable or disable displaying the Turtle's
image on the screen.
}
\details{
The Turtle must be initialized prior to using
this function, see \code{\link{turtle_init}}.

It is recommended to hide the Turtle when
performing multiple Turtle moves, for efficiency reasons,
see also \code{\link{turtle_do}}.
}
\examples{
turtle_init()
turtle_forward(4)
turtle_hide()
turtle_left(30)
turtle_forward(3)
}
\seealso{
Other TurtleGraphics: \code{\link{TurtleGraphics-package}};
  \code{\link{turtle_backward}},
  \code{\link{turtle_forward}}, \code{\link{turtle_move}};
  \code{\link{turtle_col}}, \code{\link{turtle_lty}},
  \code{\link{turtle_lwd}}, \code{\link{turtle_param}};
  \code{\link{turtle_down}}, \code{\link{turtle_up}};
  \code{\link{turtle_do}}; \code{\link{turtle_getangle}},
  \code{\link{turtle_getpos}}; \code{\link{turtle_goto}},
  \code{\link{turtle_setangle}},
  \code{\link{turtle_setpos}}; \code{\link{turtle_init}};
  \code{\link{turtle_left}}, \code{\link{turtle_right}},
  \code{\link{turtle_turn}}; \code{\link{turtle_reset}};
  \code{\link{turtle_status}}
}

