\name{turtle_status}
\alias{turtle_status}
\title{Turtle Position Status}
\usage{
turtle_status()
}
\description{
\code{turtle_status()} gives information about current
Turtle position and values of graphical parameters.
}
\details{
\code{turtle_status()} returns two element list describing
graphical parameters valus and Turtle current position.
}

