% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/status.R
\name{turtle_status}
\alias{turtle_status}
\title{Read the Turtle's Status}
\usage{
turtle_status()
}
\value{
Returns a list with three elements.
}
\description{
This function gives information about the current
Turtle's position, direction, and on display options.
}
\details{
The Turtle must be initialized prior to using
this function, see \code{\link{turtle_init}}.
}
\seealso{
Other TurtleGraphics: \code{\link{TurtleGraphics-package}},
  \code{\link{turtle_do}}, \code{\link{turtle_getpos}},
  \code{\link{turtle_goto}}, \code{\link{turtle_init}},
  \code{\link{turtle_move}}, \code{\link{turtle_param}},
  \code{\link{turtle_reset}}, \code{\link{turtle_show}},
  \code{\link{turtle_turn}}, \code{\link{turtle_up}}
}
