% Generated by roxygen2 (4.0.0): do not edit by hand
\name{turtle_status}
\alias{turtle_status}
\title{Read the Turtle's Status}
\usage{
turtle_status()
}
\value{
Returns a list with three elements.
}
\description{
This function gives information about the current
Turtle's position, direction, and on display options.
}
\details{
The Turtle must be initialized prior to using
this function, see \code{\link{turtle_init}}.
}
\seealso{
Other TurtleGraphics: \code{\link{TurtleGraphics-package}};
  \code{\link{turtle_backward}},
  \code{\link{turtle_forward}}, \code{\link{turtle_move}};
  \code{\link{turtle_col}}, \code{\link{turtle_lty}},
  \code{\link{turtle_lwd}}, \code{\link{turtle_param}};
  \code{\link{turtle_down}}, \code{\link{turtle_up}};
  \code{\link{turtle_do}}; \code{\link{turtle_getangle}},
  \code{\link{turtle_getpos}}; \code{\link{turtle_goto}},
  \code{\link{turtle_setangle}},
  \code{\link{turtle_setpos}}; \code{\link{turtle_hide}},
  \code{\link{turtle_show}}; \code{\link{turtle_init}};
  \code{\link{turtle_left}}, \code{\link{turtle_right}},
  \code{\link{turtle_turn}}; \code{\link{turtle_reset}}
}

