% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/param.R
\name{turtle_up}
\alias{turtle_up}
\alias{turtle_down}
\title{Turn on or off Turtle Trace Drawing}
\usage{
turtle_up()

turtle_down()
}
\description{
When the Turtle moves, it may or may not
leave a visible trace. These functions
control such a behavior.
}
\details{
The Turtle must be initialized prior to using
this function, see \code{\link{turtle_init}}.
}
\seealso{
Other TurtleGraphics: \code{\link{TurtleGraphics-package}},
  \code{\link{turtle_do}}, \code{\link{turtle_getpos}},
  \code{\link{turtle_goto}}, \code{\link{turtle_init}},
  \code{\link{turtle_move}}, \code{\link{turtle_param}},
  \code{\link{turtle_reset}}, \code{\link{turtle_show}},
  \code{\link{turtle_status}}, \code{\link{turtle_turn}}
}
