\docType{package}
\name{TurtleGraphics-package}
\alias{TurtleGraphics-package}
\title{Turtle Graphics in R}
\description{
TO BE DONE
}
\details{
TO BE DONE
}
\author{
Anna Cena \email{cena@rexamine.com} [aut]\cr Marek
Gagolewski \email{gagolews@rexamine.com} [aut]\cr Marcin
Kosinski \email{m.p.kosinski@gmail.com} [aut]\cr Barbara
Zogala-Siudem \email{zogala@rexamine.com} [aut, cre]
}
\seealso{
Other TurtleGraphics: \code{\link{turtle_backward}},
\code{\link{turtle_forward}}, \code{\link{turtle_move}};
\code{\link{turtle_col}}, \code{\link{turtle_lty}},
\code{\link{turtle_lwd}}, \code{\link{turtle_param}};
\code{\link{turtle_goto}}, \code{\link{turtle_setang}},
\code{\link{turtle_setpos}}; \code{\link{turtle_hide}},
\code{\link{turtle_show}}; \code{\link{turtle_init}};
\code{\link{turtle_left}}, \code{\link{turtle_right}},
\code{\link{turtle_turn}}; \code{\link{turtle_reset}};
\code{\link{turtle_status}}
}

