% Generated by roxygen2 (4.0.0): do not edit by hand
\docType{package}
\name{TurtleGraphics-package}
\alias{TurtleGraphics-package}
\title{Turtle Graphics in R}
\description{
Learn computer programming while having a jolly time
}
\details{
Move the Turtle with commands that are relative
to its own position, e.g. "move forward 100 pixels" or
"turn right 30 degrees". From these building blocks you can build
more complex shapes like circles, fractals, etc.
Combined with R control flow, functions, and recursion,
the idea of Turtle graphics allows students to get familiar with
computer programming in a very accessible and pleasant way.
}
\author{
Anna Cena [aut]\cr
Marek Gagolewski [aut]\cr
Marcin Kosinski [aut]\cr
Natalia Potocka [aut]
Barbara Zogala-Siudem [aut, cre]
}
\seealso{
Other TurtleGraphics: \code{\link{turtle_backward}},
  \code{\link{turtle_forward}}, \code{\link{turtle_move}};
  \code{\link{turtle_col}}, \code{\link{turtle_lty}},
  \code{\link{turtle_lwd}}, \code{\link{turtle_param}};
  \code{\link{turtle_down}}, \code{\link{turtle_up}};
  \code{\link{turtle_do}}; \code{\link{turtle_getangle}},
  \code{\link{turtle_getpos}}; \code{\link{turtle_goto}},
  \code{\link{turtle_setangle}},
  \code{\link{turtle_setpos}}; \code{\link{turtle_hide}},
  \code{\link{turtle_show}}; \code{\link{turtle_init}};
  \code{\link{turtle_left}}, \code{\link{turtle_right}},
  \code{\link{turtle_turn}}; \code{\link{turtle_reset}};
  \code{\link{turtle_status}}
}

